mac: make sure we only run on mac_os_x

+++ cookbooks/mac/recipes/default.rb
@@ -1,17 +1,19 @@
-execute "chflags nohidden #{node[:homedir]}/Library"
+if platform?("mac_os_x")
+  execute "chflags nohidden #{node[:homedir]}/Library"
 
-include_recipe "mac::global"
-include_recipe "mac::dock"
-include_recipe "mac::finder"
+  include_recipe "mac::global"
+  include_recipe "mac::dock"
+  include_recipe "mac::finder"
 
-mac_userdefaults "enable safari debug menu" do
-  domain "com.apple.Safari"
-  key "IncludeInternalDebugMenu"
-  value true
-  notifies :run, "execute[kill-mac-procs]"
-end
+  mac_userdefaults "enable safari debug menu" do
+    domain "com.apple.Safari"
+    key "IncludeInternalDebugMenu"
+    value true
+    notifies :run, "execute[kill-mac-procs]"
+  end
 
-execute "kill-mac-procs" do
-  command "killall Dock Finder SystemUIServer"
-  action :nothing
+  execute "kill-mac-procs" do
+    command "killall Dock Finder SystemUIServer"
+    action :nothing
+  end
 end
